2002-05-02  Ravi Pratap  <ravi@ximian.com>
	* test-113.cs : Another test for overload resolution ;-)

svn path=/trunk/mcs/; revision=4230

+++ mcs/tests/test-104.cs
@@ -5,7 +5,7 @@ using System;
 
 class Test {
         public static int Main(String[] args) {
-                long iTest = 1;
+                int iTest = 1;
 
                 System.Threading.Interlocked.Increment(ref iTest);
